Split view.rs into multiple modules
