+++ src/view/aggregation.rs
@@ -1,0 +1,197 @@
+use crate::{
+    DataType,
+    counter::{Counter, CounterMap},
+};
+use std::collections::{BTreeMap, HashMap};
+
+/// A helper trait to check if a summary data structure is empty.
+pub trait IsEmpty {
+    fn is_empty(&self) -> bool;
+}
+
+impl IsEmpty for Counter {
+    fn is_empty(&self) -> bool {
+        self.total() == 0
+    }
+}
+
+impl IsEmpty for CounterMap {
+    fn is_empty(&self) -> bool {
+        self.iter().all(|(_, counter)| counter.is_empty())
+    }
+}
+
+impl<K, V> IsEmpty for HashMap<K, V> {
+    fn is_empty(&self) -> bool {
+        HashMap::is_empty(self)
+    }
+}
+
+pub struct AggregatedData {
+    pub grouped: BTreeMap<String, BTreeMap<DataType, Counter>>,
+    pub total_block_entity: Counter,
+    pub total_entity: Counter,
+    pub total_player_data: Counter,
+    pub total_combined: Counter,
+}
+
+impl AggregatedData {
+    pub fn new(counter_map: &CounterMap) -> Self {
+        let mut grouped: BTreeMap<String, BTreeMap<DataType, Counter>> = BTreeMap::new();
+        let mut total_block_entity = Counter::new();
+        let mut total_entity = Counter::new();
+        let mut total_player_data = Counter::new();
+        let mut total_combined = Counter::new();
+
+        for (scope, counter) in counter_map.iter() {
+            grouped
+                .entry(scope.dimension.clone())
+                .or_default()
+                .entry(scope.data_type.clone())
+                .or_default()
+                .merge(counter);
+
+            match scope.data_type {
+                DataType::BlockEntity => total_block_entity.merge(counter),
+                DataType::Entity => total_entity.merge(counter),
+                DataType::Player => total_player_data.merge(counter),
+            }
+            total_combined.merge(counter);
+        }
+
+        Self {
+            grouped,
+            total_block_entity,
+            total_entity,
+            total_player_data,
+            total_combined,
+        }
+    }
+
+    fn dimension_combined(&self, dimension: &str) -> Counter {
+        let mut combined = Counter::new();
+        if let Some(types_map) = self.grouped.get(dimension) {
+            for counter in types_map.values() {
+                combined.merge(counter);
+            }
+        }
+        combined
+    }
+}
+
+/// Trait to provide summary data in a generic way for different views.
+pub trait SummaryDataProvider {
+    type ItemSummary: Clone + IsEmpty;
+
+    fn get_grouped_data(&self) -> &BTreeMap<String, BTreeMap<DataType, Self::ItemSummary>>;
+    fn get_total_block_entity_summary(&self) -> &Self::ItemSummary;
+    fn get_total_entity_summary(&self) -> &Self::ItemSummary;
+    fn get_total_player_data_summary(&self) -> &Self::ItemSummary;
+    fn get_total_combined_summary(&self) -> &Self::ItemSummary;
+    fn calculate_dimension_combined_summary(&self, dimension: &str) -> Self::ItemSummary;
+}
+
+impl SummaryDataProvider for AggregatedData {
+    type ItemSummary = Counter;
+
+    fn get_grouped_data(&self) -> &BTreeMap<String, BTreeMap<DataType, Self::ItemSummary>> {
+        &self.grouped
+    }
+    fn get_total_block_entity_summary(&self) -> &Self::ItemSummary {
+        &self.total_block_entity
+    }
+    fn get_total_entity_summary(&self) -> &Self::ItemSummary {
+        &self.total_entity
+    }
+    fn get_total_player_data_summary(&self) -> &Self::ItemSummary {
+        &self.total_player_data
+    }
+    fn get_total_combined_summary(&self) -> &Self::ItemSummary {
+        &self.total_combined
+    }
+    fn calculate_dimension_combined_summary(&self, dimension: &str) -> Self::ItemSummary {
+        self.dimension_combined(dimension)
+    }
+}
+
+pub struct AggregatedIdCountsData {
+    pub grouped: BTreeMap<String, BTreeMap<DataType, HashMap<String, u64>>>,
+    pub total_block_entity: HashMap<String, u64>,
+    pub total_entity: HashMap<String, u64>,
+    pub total_player_data: HashMap<String, u64>,
+    pub total_combined: HashMap<String, u64>,
+}
+
+impl AggregatedIdCountsData {
+    pub fn new(counter_map: &CounterMap) -> Self {
+        let mut grouped: BTreeMap<String, BTreeMap<DataType, HashMap<String, u64>>> =
+            BTreeMap::new();
+        let mut total_block_entity = HashMap::new();
+        let mut total_entity = HashMap::new();
+        let mut total_player_data = HashMap::new();
+        let mut total_combined = HashMap::new();
+
+        for (scope, counter) in counter_map.iter() {
+            let current_total_by_id = counter.total_by_id();
+            let dim_data_map = grouped
+                .entry(scope.dimension.clone())
+                .or_default()
+                .entry(scope.data_type.clone())
+                .or_default();
+
+            for (id, count) in &current_total_by_id {
+                *dim_data_map.entry(id.clone()).or_default() += *count;
+                *total_combined.entry(id.clone()).or_default() += *count;
+                match scope.data_type {
+                    DataType::BlockEntity => {
+                        *total_block_entity.entry(id.clone()).or_default() += *count
+                    }
+                    DataType::Entity => *total_entity.entry(id.clone()).or_default() += *count,
+                    DataType::Player => *total_player_data.entry(id.clone()).or_default() += *count,
+                }
+            }
+        }
+        Self {
+            grouped,
+            total_block_entity,
+            total_entity,
+            total_player_data,
+            total_combined,
+        }
+    }
+
+    fn dimension_combined(&self, dimension: &str) -> HashMap<String, u64> {
+        let mut combined = HashMap::new();
+        if let Some(types_map) = self.grouped.get(dimension) {
+            for id_map in types_map.values() {
+                for (id, count) in id_map {
+                    *combined.entry(id.clone()).or_default() += *count;
+                }
+            }
+        }
+        combined
+    }
+}
+
+impl SummaryDataProvider for AggregatedIdCountsData {
+    type ItemSummary = HashMap<String, u64>;
+
+    fn get_grouped_data(&self) -> &BTreeMap<String, BTreeMap<DataType, Self::ItemSummary>> {
+        &self.grouped
+    }
+    fn get_total_block_entity_summary(&self) -> &Self::ItemSummary {
+        &self.total_block_entity
+    }
+    fn get_total_entity_summary(&self) -> &Self::ItemSummary {
+        &self.total_entity
+    }
+    fn get_total_player_data_summary(&self) -> &Self::ItemSummary {
+        &self.total_player_data
+    }
+    fn get_total_combined_summary(&self) -> &Self::ItemSummary {
+        &self.total_combined
+    }
+    fn calculate_dimension_combined_summary(&self, dimension: &str) -> Self::ItemSummary {
+        self.dimension_combined(dimension)
+    }
+}
